Vacancy implement

+++ src/main/java/by/team34/dto/RoleDto.java
@@ -1,19 +1,19 @@
 package by.team34.dto;
 
 public class RoleDto {
-    private int id;
+    private Long id;
     private String name;
 
-    public RoleDto(int id, String name) {
+    public RoleDto(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
